feat: Implement Payroll, Purchase Order, and Supplier Contract Controllers
- Added PayrollController to manage payroll records, including CRUD operations and payroll generation.
- Introduced PurchaseOrderController for handling purchase orders with endpoints for retrieval, creation, and updates.
- Created SupplierContractController to manage supplier contracts, including status updates and retrieval by supplier.
- Added DTOs for PayrollRecord and SupplierContract to facilitate data transfer.
- Established database migration scripts for consolidating user JWT schema and ensuring necessary tables exist.
- Seeded initial data for brands, categories, suppliers, roles, users, and other entities for testing purposes.

+++ backend/src/main/java/com/smalltrend/controller/PricingRuleController.java
@@ -65,7 +65,7 @@ public class PricingRuleController {
 
     // Simple evaluation: list rules by product variant id (if provided)
     @GetMapping("/bulk/evaluate")
-    public List<BulkPricingRule> evaluate(@RequestParam(required = false) Long productVariantId) {
+    public List<BulkPricingRule> evaluate(@RequestParam(required = false) Integer productVariantId) {
         if (productVariantId == null) {
             return activeBulk();
         }
